Fix links. Remove unused deps.
pryr::partial -> purrr::partial
use rlang for standalone types-check.

devtools, ggplot2 are not used.
lintr is only a developement dep (not required for using beautier)

hunspell is a dep of spelling and not used directly.

assertive has been archived, and replaced here by rlang.

Maintainer is no longer required as the maintainer is the c("aut", "cre")